refactor: Transaction komponens használata
